refactor : MessageTextArea로 이름 변경
