+++ src/components/molecules/MessageTextArea/index.tsx
@@ -2,14 +2,14 @@ import React from 'react';
 import { TextArea } from '@/components';
 import * as S from './styles';
 
-interface MessageInputProps {
+interface MessageTextAreaProps {
   value: string;
   onChange: (e: React.ChangeEvent<HTMLTextAreaElement>) => void;
   placeholder?: string;
   error?: string;
 }
 
-const MessageInput = ({ value, onChange, placeholder, error }: MessageInputProps) => {
+const MessageTextArea = ({ value, onChange, placeholder, error }: MessageTextAreaProps) => {
   return (
     <S.Container>
       <S.Wrapper>
@@ -25,4 +25,4 @@ const MessageInput = ({ value, onChange, placeholder, error }: MessageInputProps
   );
 };
 
-export default MessageInput;
+export default MessageTextArea;
